+++ src/Raven.CodeAnalysis/Binder/TypeMemberBinder.Properties.cs
@@ -280,6 +280,14 @@ internal partial class TypeMemberBinder : Binder
         if (explicitInterfaceProperty is not null && sourcePropertySymbol is not null)
             sourcePropertySymbol.SetExplicitInterfaceImplementations(ImmutableArray.Create(explicitInterfaceProperty));
 
+        if (isPartial && sourcePropertySymbol is not null)
+        {
+            if (isPartialImplementationSyntax)
+                sourcePropertySymbol.MarkAsPartialImplementation();
+            else if (isPartialDefinitionSyntax)
+                sourcePropertySymbol.MarkAsPartialDefinition();
+        }
+
         // Bind property initializer (if any) with the property's declared type as target type.
         if (propertyDecl.Initializer is not null)
         {
@@ -1089,10 +1097,6 @@ internal partial class TypeMemberBinder : Binder
                 return binders;
             }
 
-            if (isPartialImplementationSyntax)
-                sourcePropertySymbol.MarkAsPartialImplementation();
-            else if (isPartialDefinitionSyntax)
-                sourcePropertySymbol.MarkAsPartialDefinition();
         }
 
         return binders;